add routes and pages
