--- client/src/components/DefaultCard/DefaultCard.css
@@ -1,7 +1,0 @@
-.default_card {
-    background: pink;
-    width: 60%;
-    height: 600px;
-    margin: 0 auto;
-    margin-top: 100px;
-}
