+++ client/src/components/Logo/style.css
@@ -1,0 +1,8 @@
+.logo-container {
+    width: 150px;
+}
+
+img {
+    width: 150px;
+    margin: 20px 0;
+}
