+++ package.json
@@ -28,6 +28,8 @@
     "mysql2": "^1.7.0",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
+    "react-router-dom": "^5.1.2",
+    "react-strap": "0.0.1",
     "reactstrap": "^8.1.1",
     "redux": "^4.0.4",
     "sequelize": "^5.19.7"
